Refactor the code duplacation in luaot-trampoline
