+++ src/luaot_switches.c
@@ -1,668 +1,3 @@
-/*
- * Lua bytecode-to-C compiler
- */
-
-// This luac-derived code is incompatible with lua_assert because it calls the
-// GETARG macros even for opcodes where it is not appropriate to do so.
-#undef LUAI_ASSERT
-
-#include <ctype.h>
-#include <errno.h>
-#include <stdarg.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include "lua.h"
-#include "lauxlib.h"
-
-#include "ldebug.h"
-#include "lobject.h"
-#include "lopcodes.h"
-#include "lopnames.h"
-#include "lstate.h"
-#include "lundump.h"
-
-//
-// Command-line arguments and main function
-// ----------------------------------------
-//
-// This part should not depend much on the Lua version
-//
-
-static const char *program_name    = "luaot";
-static const char *input_filename  = NULL;
-static const char *output_filename = NULL;
-static const char *module_name     = NULL;
-static FILE * output_file = NULL;
-static int nfunctions = 0;
-static TString **tmname;
-
-static
-void usage()
-{
-    fprintf(stderr, "usage: %s input.lua -o output.c\n", program_name);
-}
-
-static
-void fatal_error(const char *msg)
-{
-    fprintf(stderr, "%s: %s\n", program_name, msg);
-    exit(1);
-}
-
-static
-__attribute__ ((format (printf, 1, 2)))
-void print(const char *fmt, ...)
-{
-    va_list args;
-    va_start(args, fmt);
-    vfprintf(output_file, fmt, args);
-    va_end(args);
-}
-
-static
-__attribute__ ((format (printf, 1, 2)))
-void println(const char *fmt, ...)
-{
-    va_list args;
-    va_start(args, fmt);
-    vfprintf(output_file, fmt, args);
-    va_end(args);
-    fprintf(output_file, "\n");
-}
-
-static
-void printnl()
-{
-    // This separate function avoids Wformat-zero-length warnings with println
-    fprintf(output_file, "\n");
-}
-
-
-static void doargs(int argc, char **argv)
-{
-    // I wonder if I should just use getopt instead of parsing options by hand
-    program_name = argv[0];
-
-    int do_opts = 1;
-    int npos = 0;
-    for (int i = 1; i < argc; i++) {
-        const char *arg = argv[i];
-        if (do_opts && arg[0] == '-') {
-            if (0 == strcmp(arg, "--")) {
-                do_opts = 0;
-            } else if (0 == strcmp(arg, "-h")) {
-                usage();
-                exit(0);
-            } else if (0 == strcmp(arg, "-o")) {
-                i++;
-                if (i >= argc) { fatal_error("missing argument for -o"); }
-                output_filename = argv[i];
-            } else {
-                fprintf(stderr, "unknown option %s\n", arg);
-                exit(1);
-            }
-        } else {
-            switch (npos) {
-                case 0:
-                    input_filename = arg;
-                    break;
-                default:
-                    fatal_error("too many positional arguments");
-                    break;
-            }
-            npos++;
-        }
-    }
-
-    if (output_filename == NULL) {
-        usage();
-        exit(1);
-    }
-}
-
-static const char *get_module_name(const char *);
-static void print_functions();
-static void print_source_code();
-
-int main(int argc, char **argv)
-{
-    // Process input arguments
-
-    doargs(argc, argv);
-    module_name = get_module_name(output_filename);
-
-    // Read the input
-
-    lua_State *L = luaL_newstate();
-    if (luaL_loadfile(L, input_filename) != LUA_OK) {
-        fatal_error(lua_tostring(L,-1));
-    }
-    Proto *proto = getproto(s2v(L->top-1));
-    tmname = G(L)->tmname;
-
-    // Generate the file
-
-    output_file = fopen(output_filename, "w");
-    if (output_file == NULL) { fatal_error(strerror(errno)); }
-
-    println("#include \"trampoline_header.c\"");
-    printnl();
-    print_functions(proto);
-    printnl();
-    print_source_code();
-    printnl();
-    println("#define LUAOT_LUAOPEN_NAME luaopen_%s", module_name);
-    printnl();
-    println("#include \"trampoline_footer.c\"");
-}
-
-// Deduce the Lua module name given the file name
-// Example:  ./foo/bar/baz.c -> foo_bar_baz
-static
-const char *get_module_name(const char *filename)
-{
-    size_t n = strlen(filename);
-
-    int has_extension = 0;
-    size_t sep = 0;
-    for (size_t i = 0; i < n; i++) {
-        if (filename[i] == '.') {
-            has_extension = 1;
-            sep = i;
-        }
-    }
-
-    if (!has_extension || 0 != strcmp(filename + sep, ".c")) {
-        fatal_error("output file does not have a \".c\" extension");
-    }
-
-    char *module_name = malloc(sep+1);
-    for (size_t i = 0; i < sep; i++) {
-        int c = filename[i];
-        if (c == '/' || c == '.') {
-            module_name[i] = '_';
-        } else {
-            module_name[i] = c;
-        }
-    }
-    module_name[sep] = '\0';
-
-
-    for (size_t i = 0; i < sep; i++) {
-        int c = module_name[i];
-        if (!isalnum(c) && c != '_') {
-            fatal_error("output module name contains invalid characters");
-        }
-    }
-
-    return module_name;
-}
-
-//
-// Printing bytecode information
-// -----------------------------
-//
-// These functions are copied from luac.c (and reindented)
-//
-
-#define UPVALNAME(x) ((f->upvalues[x].name) ? getstr(f->upvalues[x].name) : "-")
-#define VOID(p) ((const void*)(p))
-#define eventname(i) (getstr(tmname[i]))
-
-static
-void PrintString(const TString* ts)
-{
-    const char* s = getstr(ts);
-    size_t i,n = tsslen(ts);
-    print("\"");
-    for (i=0; i<n; i++) {
-        int c=(int)(unsigned char)s[i];
-        switch (c) {
-            case '"':
-                print("\\\"");
-                break;
-            case '\\':
-                print("\\\\");
-                break;
-            case '\a':
-                print("\\a");
-                break;
-            case '\b':
-                print("\\b");
-                break;
-            case '\f':
-                print("\\f");
-                break;
-            case '\n':
-                print("\\n");
-                break;
-            case '\r':
-                print("\\r");
-                break;
-            case '\t':
-                print("\\t");
-                break;
-            case '\v':
-                print("\\v");
-                break;
-            default:
-                if (isprint(c)) {
-                    print("%c",c);
-                } else {
-                    print("\\%03d",c);
-                }
-                break;
-        }
-    }
-    print("\"");
-}
-
-#if 0
-static
-void PrintType(const Proto* f, int i)
-{
-    const TValue* o=&f->k[i];
-    switch (ttypetag(o)) {
-        case LUA_VNIL:
-            printf("N");
-            break;
-        case LUA_VFALSE:
-        case LUA_VTRUE:
-            printf("B");
-            break;
-        case LUA_VNUMFLT:
-            printf("F");
-            break;
-        case LUA_VNUMINT:
-            printf("I");
-            break;
-        case LUA_VSHRSTR:
-        case LUA_VLNGSTR:
-            printf("S");
-            break;
-        default: /* cannot happen */
-            printf("?%d",ttypetag(o));
-            break;
-    }
-    printf("\t");
-}
-#endif
-
-static
-void PrintConstant(const Proto* f, int i)
-{
-    const TValue* o=&f->k[i];
-    switch (ttypetag(o)) {
-        case LUA_VNIL:
-            print("nil");
-            break;
-        case LUA_VFALSE:
-            printf("false");
-            break;
-        case LUA_VTRUE:
-            printf("true");
-            break;
-        case LUA_VNUMFLT:
-            {
-                char buff[100];
-                sprintf(buff,LUA_NUMBER_FMT,fltvalue(o));
-                print("%s",buff);
-                if (buff[strspn(buff,"-0123456789")]=='\0') print(".0");
-                break;
-            }
-        case LUA_VNUMINT:
-            print(LUA_INTEGER_FMT, ivalue(o));
-            break;
-        case LUA_VSHRSTR:
-        case LUA_VLNGSTR:
-            PrintString(tsvalue(o));
-            break;
-        default: /* cannot happen */
-            print("?%d",ttypetag(o));
-            break;
-    }
-}
-
-#define COMMENT		"\t; "
-#define EXTRAARG	GETARG_Ax(code[pc+1])
-#define EXTRAARGC	(EXTRAARG*(MAXARG_C+1))
-#define ISK		(isk ? "k" : "")
-
-static
-void luaot_PrintOpcodeComment(Proto *f, int pc)
-{
-    // Adapted from the PrintCode function of luac.c
-    const Instruction *code = f->code;
-    const Instruction i = code[pc];
-    OpCode o = GET_OPCODE(i);
-    int a=GETARG_A(i);
-    int b=GETARG_B(i);
-    int c=GETARG_C(i);
-    int ax=GETARG_Ax(i);
-    int bx=GETARG_Bx(i);
-    int sb=GETARG_sB(i);
-    int sc=GETARG_sC(i);
-    int sbx=GETARG_sBx(i);
-    int isk=GETARG_k(i);
-    int line=luaG_getfuncline(f,pc);
-
-    print("      //");
-    print(" %d\t", pc);
-    if (line > 0) {
-        print("[%d]\t", line);
-    } else {
-        print("[-]\t");
-    }
-    print("%-9s\t", opnames[o]);
-    switch (o) {
-        case OP_MOVE:
-            print("%d %d",a,b);
-            break;
-        case OP_LOADI:
-            print("%d %d",a,sbx);
-            break;
-        case OP_LOADF:
-            print("%d %d",a,sbx);
-            break;
-        case OP_LOADK:
-            print("%d %d",a,bx);
-            print(COMMENT); PrintConstant(f,bx);
-            break;
-        case OP_LOADKX:
-            print("%d",a);
-            print(COMMENT); PrintConstant(f,EXTRAARG);
-            break;
-        case OP_LOADFALSE:
-            print("%d",a);
-            break;
-        case OP_LFALSESKIP:
-            print("%d",a);
-            break;
-        case OP_LOADTRUE:
-            print("%d",a);
-            break;
-        case OP_LOADNIL:
-            print("%d %d",a,b);
-            print(COMMENT "%d out",b+1);
-            break;
-        case OP_GETUPVAL:
-            print("%d %d",a,b);
-            print(COMMENT "%s", UPVALNAME(b));
-            break;
-        case OP_SETUPVAL:
-            print("%d %d",a,b);
-            print(COMMENT "%s", UPVALNAME(b));
-            break;
-        case OP_GETTABUP:
-            print("%d %d %d",a,b,c);
-            print(COMMENT "%s", UPVALNAME(b));
-            print(" "); PrintConstant(f,c);
-            break;
-        case OP_GETTABLE:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_GETI:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_GETFIELD:
-            print("%d %d %d",a,b,c);
-            print(COMMENT); PrintConstant(f,c);
-            break;
-        case OP_SETTABUP:
-            print("%d %d %d%s",a,b,c,ISK);
-            print(COMMENT "%s",UPVALNAME(a));
-            print(" "); PrintConstant(f,b);
-            if (isk) { print(" "); PrintConstant(f,c); }
-            break;
-        case OP_SETTABLE:
-            print("%d %d %d%s",a,b,c,ISK);
-            if (isk) { print(COMMENT); PrintConstant(f,c); }
-            break;
-        case OP_SETI:
-            print("%d %d %d%s",a,b,c,ISK);
-            if (isk) { print(COMMENT); PrintConstant(f,c); }
-            break;
-        case OP_SETFIELD:
-            print("%d %d %d%s",a,b,c,ISK);
-            print(COMMENT); PrintConstant(f,b);
-            if (isk) { print(" "); PrintConstant(f,c); }
-            break;
-        case OP_NEWTABLE:
-            print("%d %d %d",a,b,c);
-            print(COMMENT "%d",c+EXTRAARGC);
-            break;
-        case OP_SELF:
-            print("%d %d %d%s",a,b,c,ISK);
-            if (isk) { print(COMMENT); PrintConstant(f,c); }
-            break;
-        case OP_ADDI:
-            print("%d %d %d",a,b,sc);
-            break;
-        case OP_ADDK:
-            print("%d %d %d",a,b,c);
-            print(COMMENT); PrintConstant(f,c);
-            break;
-        case OP_SUBK:
-            print("%d %d %d",a,b,c);
-            print(COMMENT); PrintConstant(f,c);
-            break;
-        case OP_MULK:
-            print("%d %d %d",a,b,c);
-            print(COMMENT); PrintConstant(f,c);
-            break;
-        case OP_MODK:
-            print("%d %d %d",a,b,c);
-            print(COMMENT); PrintConstant(f,c);
-            break;
-        case OP_POWK:
-            print("%d %d %d",a,b,c);
-            print(COMMENT); PrintConstant(f,c);
-            break;
-        case OP_DIVK:
-            print("%d %d %d",a,b,c);
-            print(COMMENT); PrintConstant(f,c);
-            break;
-        case OP_IDIVK:
-            print("%d %d %d",a,b,c);
-            print(COMMENT); PrintConstant(f,c);
-            break;
-        case OP_BANDK:
-            print("%d %d %d",a,b,c);
-            print(COMMENT); PrintConstant(f,c);
-            break;
-        case OP_BORK:
-            print("%d %d %d",a,b,c);
-            print(COMMENT); PrintConstant(f,c);
-            break;
-        case OP_BXORK:
-            print("%d %d %d",a,b,c);
-            print(COMMENT); PrintConstant(f,c);
-            break;
-        case OP_SHRI:
-            print("%d %d %d",a,b,sc);
-            break;
-        case OP_SHLI:
-            print("%d %d %d",a,b,sc);
-            break;
-        case OP_ADD:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_SUB:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_MUL:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_MOD:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_POW:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_DIV:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_IDIV:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_BAND:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_BOR:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_BXOR:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_SHL:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_SHR:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_MMBIN:
-            print("%d %d %d",a,b,c);
-            print(COMMENT "%s",eventname(c));
-            break;
-        case OP_MMBINI:
-            print("%d %d %d %d",a,sb,c,isk);
-            print(COMMENT "%s",eventname(c));
-            if (isk) print(" flip");
-            break;
-        case OP_MMBINK:
-            print("%d %d %d %d",a,b,c,isk);
-            print(COMMENT "%s ",eventname(c)); PrintConstant(f,b);
-            if (isk) print(" flip");
-            break;
-        case OP_UNM:
-            print("%d %d",a,b);
-            break;
-        case OP_BNOT:
-            print("%d %d",a,b);
-            break;
-        case OP_NOT:
-            print("%d %d",a,b);
-            break;
-        case OP_LEN:
-            print("%d %d",a,b);
-            break;
-        case OP_CONCAT:
-            print("%d %d",a,b);
-            break;
-        case OP_CLOSE:
-            print("%d",a);
-            break;
-        case OP_TBC:
-            print("%d",a);
-            break;
-        case OP_JMP:
-            print("%d",GETARG_sJ(i));
-            print(COMMENT "to %d",GETARG_sJ(i)+pc+2);
-            break;
-        case OP_EQ:
-            print("%d %d %d",a,b,isk);
-            break;
-        case OP_LT:
-            print("%d %d %d",a,b,isk);
-            break;
-        case OP_LE:
-            print("%d %d %d",a,b,isk);
-            break;
-        case OP_EQK:
-            print("%d %d %d",a,b,isk);
-            print(COMMENT); PrintConstant(f,b);
-            break;
-        case OP_EQI:
-            print("%d %d %d",a,sb,isk);
-            break;
-        case OP_LTI:
-            print("%d %d %d",a,sb,isk);
-            break;
-        case OP_LEI:
-            print("%d %d %d",a,sb,isk);
-            break;
-        case OP_GTI:
-            print("%d %d %d",a,sb,isk);
-            break;
-        case OP_GEI:
-            print("%d %d %d",a,sb,isk);
-            break;
-        case OP_TEST:
-            print("%d %d",a,isk);
-            break;
-        case OP_TESTSET:
-            print("%d %d %d",a,b,isk);
-            break;
-        case OP_CALL:
-            print("%d %d %d",a,b,c);
-            print(COMMENT);
-            if (b==0) print("all in "); else print("%d in ",b-1);
-            if (c==0) print("all out"); else print("%d out",c-1);
-            break;
-        case OP_TAILCALL:
-            print("%d %d %d",a,b,c);
-            print(COMMENT "%d in",b-1);
-            break;
-        case OP_RETURN:
-            print("%d %d %d",a,b,c);
-            print(COMMENT);
-            if (b==0) print("all out"); else print("%d out",b-1);
-            break;
-        case OP_RETURN0:
-            break;
-        case OP_RETURN1:
-            print("%d",a);
-            break;
-        case OP_FORLOOP:
-            print("%d %d",a,bx);
-            print(COMMENT "to %d",pc-bx+2);
-            break;
-        case OP_FORPREP:
-            print("%d %d",a,bx);
-            print(COMMENT "to %d",pc+bx+2);
-            break;
-        case OP_TFORPREP:
-            print("%d %d",a,bx);
-            print(COMMENT "to %d",pc+bx+2);
-            break;
-        case OP_TFORCALL:
-            print("%d %d",a,c);
-            break;
-        case OP_TFORLOOP:
-            print("%d %d",a,bx);
-            print(COMMENT "to %d",pc-bx+2);
-            break;
-        case OP_SETLIST:
-            print("%d %d %d",a,b,c);
-            break;
-        case OP_CLOSURE:
-            print("%d %d",a,bx);
-            print(COMMENT "%p",VOID(f->p[bx]));
-            break;
-        case OP_VARARG:
-            print("%d %d",a,c);
-            print(COMMENT);
-            if (c==0) print("all out"); else print("%d out",c-1);
-            break;
-        case OP_VARARGPREP:
-            print("%d",a);
-            break;
-        case OP_EXTRAARG:
-            print("%d",ax);
-            break;
-#if 0
-        default:
-            print("%d %d %d",a,b,c);
-            print(COMMENT "not handled");
-            break;
-#endif
-    }
-    print("\n");
-}
-
-
 //
 // The compiler part of the compiler
 // ---------------------------------
@@ -1545,70 +880,4 @@ void create_function(Proto *f)
     println("  }");
     println("}");
     printnl();
-}
-
-static
-void create_functions(Proto *p)
-{
-    // luaot_footer.c should use the same traversal order as this.
-    create_function(p);
-    for (int i = 0; i < p->sizep; i++) {
-        create_functions(p->p[i]);
-    }
-}
-
-static
-void print_functions(Proto *p)
-{
-    create_functions(p);
-
-    println("static AotCompiledFunction LUAOT_FUNCTIONS[] = {");
-    for (int i = 0; i < nfunctions; i++) {
-        println("  magic_implementation_%02d,", i);
-    }
-    println("  NULL");
-    println("};");
-}
-
-static
-void print_source_code()
-{
-    // Since the code we are generating is lifted from lvm.c, we need it to use
-    // Lua functions instead of C functions. And to create the Lua functions,
-    // we have to `load` them from source code.
-    //
-    // The most readable approach would be to bundle this Lua source code as a
-    // big C string literal. However, C compilers have limits on how big a
-    // string literal can be, so instead of using a string literal, we use a
-    // plain char array instead.
-
-    FILE *infile = fopen(input_filename, "r");
-    if (!infile) { fatal_error("could not open input file a second time"); }
-
-    println("static const char LUAOT_MODULE_SOURCE_CODE[] = {");
-
-    int c;
-    int col = 0;
-    do {
-        if (col == 0) {
-            print(" ");
-        }
-
-        c = fgetc(infile);
-        if (c == EOF) {
-            print(" %3d", 0);
-        } else {
-            print(" %3d", c);
-            print(",");
-        }
-
-        col++;
-        if (col == 16 || c == EOF) {
-            print("\n");
-            col = 0;
-        }
-    } while (c != EOF);
-    println("};");
-
-    fclose(infile);
 }
